move skills out
