+++ src/directives/lang/language.directive.ts
@@ -1,5 +1,5 @@
 import {Directive, OnInit} from '@angular/core';
-import {LevelComponent} from "../../app/main-view/experience/skills/level/level.component";
+import {LevelComponent} from "../../app/main-view/skills/level/level.component";
 
 @Directive({
   selector: '[pLanguage]'
